SpreadsheetSelectionMenuTest moved tests from SpreadsheetSelectionMenuPatternTest
